chore: instrument the actual method calling `response.text` or `response.json` and parse the results to avoid it getting called more than once in an instrumented promise chain
