+++ lib/subscribers/openai/config.js
@@ -68,20 +68,22 @@ module.exports = {
       ]
     },
     {
-      path: './openai/api-promise-constructor.js',
+      path: './openai/parse-response.js',
       instrumentations: [
         {
-          channelName: 'nr_apiPromise',
+          channelName: 'nr_parseResponse',
           module: { name: 'openai', versionRange: '>=4.87.0 <5.0.0', filePath: 'core.js' },
           functionQuery: {
-            className: 'APIPromise'
+            functionName: 'defaultParseResponse',
+            kind: 'Async'
           }
         },
         {
-          channelName: 'nr_apiPromise',
-          module: { name: 'openai', versionRange: '>=5.0.0', filePath: 'core/api-promise.js' },
+          channelName: 'nr_parseResponse',
+          module: { name: 'openai', versionRange: '>=5.0.0', filePath: 'internal/parse.js' },
           functionQuery: {
-            className: 'APIPromise'
+            functionName: 'defaultParseResponse',
+            kind: 'Async'
           }
         }
       ]
